Reland r145993.
Original codereview at http://codereview.chromium.org/10766016/

[Sync] Add location parameter to SyncChange in Sync API.

SyncChanges now require the location they were created as a parameter. This
is then used in the case that the change violates and invariant. We display
the location and type of erroneous change in about:sync, in addition to
disabling the offending datatype.

TBR=stevet@chromium.org, rlarocque@chromium.org
BUG=135108
TEST=unit_tests


Review URL: https://chromiumcodereview.appspot.com/10690128

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@146050 0039d316-1c4b-4281-b951-d872f2087c98