Make <browser> tag shim handle tags created after load.
The script gets injected and run so early (as the script context is created)
that the mutation observer would not get registered correctly. The fix is to
move the mutation observer creation into the DOMContentLoaded event handler (it
can't observe mutations before that point anyway).

BUG=136856
R=asargent@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10704174

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@146512 0039d316-1c4b-4281-b951-d872f2087c98