Rename APP_EXITING notification to ALL_BROWSERS_CLOSING.
With packaged apps the action of exiting Chrome (e.g. via the wrench menu)
will not exit the browser process, so this notification is being renamed to
make that clearer. Also it seems to be misused currently by the multi-profile
code, which has been updated to handle an exit of this type being aborted
by an onbeforeload handler.

BUG=130457
TEST=Existing tests

Review URL: https://chromiumcodereview.appspot.com/10692094

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@146556 0039d316-1c4b-4281-b951-d872f2087c98