gdata: Simplify Search* tests in GDataFileSystemTest
Previously, these tests were using GetGDataEntryByPath(). This is not good
as it's a private function, which should be eliminated as part of migration
to the new metadata storage.

For Search* tests, it should be sufficient to check the resource ID of the
returned entries. After all, GetEntryInfoAndCompare and friends were just
checking the resource ID.

BUG=137234
TEST=out/Release/unit_tests --gtest_filter=GData*

Review URL: https://chromiumcodereview.appspot.com/10736070

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@146664 0039d316-1c4b-4281-b951-d872f2087c98