PyAuto: Fix race in wait for login form reload after policy refresh
Refreshing device policy with the login form visible causes the form to
reload. It seems that on some of the autotest bots, the reload is already
finished by the time _WaitForLoginFormToReload() is called, causing the
wait to miss the reload and time out eventually.

This CL introduces a sentinel variable that gets updated when the form
reloads, allowing _WaitForLoginFormToReload() to reliably check for a
reload, whether it happened before the function was called or during its
runtime.

BUG=None (bugtracker currently down)
TEST=PyAuto policy tests on Chrome OS


Review URL: https://chromiumcodereview.appspot.com/10790049

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@147418 0039d316-1c4b-4281-b951-d872f2087c98