Prevent loading libffmpegsumo.so to python (autotests) process.
This is needed for ASAN environment because libffmpegsumo.so, being
built with ASAN support, fails to load to python process (which is not
ASAN-ed currently). At the same time it loads file to chrome or to
gtest processes (like content_unittests) because they do have ASAN
runtime.

FYI: Alternative approach (https://chromiumcodereview.appspot.com/10692190/)
was skipped because it disallowed ASAN to test libffmpegsumo.so.

BUG=chromium-os:32259
TESTS=unit


Review URL: https://chromiumcodereview.appspot.com/10786038

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@148183 0039d316-1c4b-4281-b951-d872f2087c98