Move printing.* prefs from local state to profile
This CL moves the printing.* prefs to the profile where they arguably
should have resided in the first place. Unit tests are simplified by
this change as the prefs can now be checked without verifying that local
state exists. There is no user-visible change as the prefs are not
exposed via the UI.

BUG=137814
TEST=all unit_tests pass, all printing browser_tests pass


Review URL: https://chromiumcodereview.appspot.com/10808006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@148531 0039d316-1c4b-4281-b951-d872f2087c98