Turn on the fixed position stacking context flag by default
Because there's an expectation that this flag will need to be reverted on a
branch, add both the enable and the disable flags so that only
web_contents_impl.cc will need to be touched to change the default value.
Both can be removed later, once this setting sticks.

BUG=none
TEST=none


Review URL: https://chromiumcodereview.appspot.com/10825027

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@148572 0039d316-1c4b-4281-b951-d872f2087c98