Pyauto perf test testNewTabDocs now waits for a different page element.
testNewTabDocs loads Google Docs and then waits until an expected
substring appears on the page before continuing.  Previously, the test
waited for a string that only appears when viewing the "my drive" folder,
because it was assumed that we'd always see that folder when opening google
docs.

However, sometimes the "shared with me" folder appears on page load.  This
appears to be because google docs opens the last folder opened by the
account.  Since we have another perf test that clicks on the "shared with me"
folder, then the current test will sometimes fail to realize the page loaded,
since an unexpected screen is displayed after page load.

The fix is to have this test wait for a different substring on the page,
which appears regardless of which folder shows up when loading google docs.

BUG=None
TEST=Verified the test runs for 10 iterations whereas previously it failed
after only 1 or 2 iterations.
NOTRY=true


Review URL: https://chromiumcodereview.appspot.com/10832127

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@149777 0039d316-1c4b-4281-b951-d872f2087c98