Refactoring: Extract gesture event filtering into helper class.
In the course of writing
http://src.chromium.org/viewvc/chrome?view=rev&revision=150463
I noticed that there was getting to be quite a bit of gesture
event related code in RenderWidgetHostImpl. This change gathers
this code together in an external helper class so that some
restructuring needed prior to completing 137555 is easier. This
CL should not change any functionality -- only moves code.

BUG=137555
TBR=jam@chromium.org


Review URL: https://chromiumcodereview.appspot.com/10823262

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@151063 0039d316-1c4b-4281-b951-d872f2087c98