Fixing bug in updating file list in save-as dialog.
Problem occuired because file manager code uses the selectedIndexes property of selectionModel. But setter for this methos is implemented only for multi-selection list, but not fo single selection (which is used in the save-as dialog). It fails to restore selection and consequently fails to restore lead item. Resetting lead item stops renaming.

BUG=138975
TEST=Manual test


Review URL: https://chromiumcodereview.appspot.com/10823201

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@151200 0039d316-1c4b-4281-b951-d872f2087c98