Downgrade media::Pipeline CHECKs into NOTREACHED() + no-op.
r150930 set Pipeline::running_ to false when an error is encountered as opposed to keeping the Pipeline running. The end result is that an in-flight Seek() can hit CHECK(running_). We want a no-op in that situation in release mode, not a process crash.

BUG=141923
TBR=fischman

Review URL: https://chromiumcodereview.appspot.com/10855167

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@151662 0039d316-1c4b-4281-b951-d872f2087c98