Refactor PPAPI browser tests and the new NaCl smoke tests to share code.
The shared code monitors and decodes an incoming stream of automation events.
The shared code is mostly identical to code used by the NaCl smoke tests, but
extended to handle cases where the test is reloaded and a case where the same
observer was reused multiple times.

This change eliminates a timer from the PPAPI tests that was a leftover from
when they were ui_tests.  Before this change, the internal browser_test timeout
would always preempt this timer.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=2959
TEST= browser_tests --gtest_filter PPAPI*


Review URL: https://chromiumcodereview.appspot.com/10860002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@152854 0039d316-1c4b-4281-b951-d872f2087c98