Revert 139366 - [OSX/Pepper] Don't have browser cache shmem ref behind ImageData.
On OSX the browser has to create shmem blocks for sandboxed processes.
The browser process can keep a handle for synchronizing between
processes by id, but pepper's implementation just passes the handle
between processes so that is not necessary.

Before, the browser ran out of handles which breaks all sorts of stuff, but everything kept running, with various visual glitches (and no fds in browser is really bad).  With this change, the renderer crashes.  That sounds bad, but is actually better in terms of resolving the underlying problems.

Also, histogram sizes for all platforms.  Using a TransportDIB for
small images is probably a bad idea.  The histograms should show how
things are being used in the wild.

BUG=129430,141541
TEST=flapper works, histograms appear.


Review URL: https://chromiumcodereview.appspot.com/10424007

TBR=shess@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10879050

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@153136 0039d316-1c4b-4281-b951-d872f2087c98